Address reviewer feedback: use DurationProvider, make eventReporter non-null, remove FIELD_SOURCE
Committed-By-Agent: claude
