+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/addresselement/StripeHostedPlacesClientProxy.kt
@@ -9,7 +9,6 @@ import com.stripe.android.ui.core.elements.autocomplete.model.AutocompletePredic
 import com.stripe.android.ui.core.elements.autocomplete.model.FindAutocompletePredictionsResponse
 import java.util.Locale
 import java.util.UUID
-import kotlin.time.TimeSource
 
 internal class StripeHostedPlacesClientProxy(
     private val repository: StripeAutocompleteRepository,
@@ -46,7 +45,7 @@ internal class StripeHostedPlacesClientProxy(
             lastQueryLength = q.length
             sessionToken
         }
-        val fetchStart = TimeSource.Monotonic.markNow()
+        eventReporter.onAutocompleteFetchStarted(token)
         return repository.findAutocompletePredictions(
             query = q,
             country = country,
@@ -70,7 +69,6 @@ internal class StripeHostedPlacesClientProxy(
             eventReporter.onAutocompleteSuggestionsReturned(
                 sessionToken = token,
                 resultCount = response.autocompletePredictions.size,
-                fetchDuration = fetchStart.elapsedNow(),
             )
         }.onFailure { error ->
             eventReporter.onAutocompleteError(sessionToken = token, error = error)
