+++ stripe-core/src/main/java/com/stripe/android/core/utils/DurationProvider.kt
@@ -50,6 +50,7 @@ interface DurationProvider {
         IntentConfirmationChallengeWebViewLoaded,
         PaymentMethodMessaging,
         AddressAutocompleteSession,
+        AddressAutocompleteFetch,
         AddressElementCompletion
     }
 }
